requirements.txt
asgiref==3.3.1
Django==3.1.4
django-rest-framework==0.1.0
django-spurl==0.6.7
django-taggit==1.3.0
djangorestframework==3.12.2
Pillow==8.0.1
psycopg2-binary==2.8.6
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
URLObject==2.4.3
gunicorn==20.0.4
